5.14. Модуляция
- разделение на модули
- удаление лишнего кода
